+++ lib/l10n/app_localizations.dart
@@ -3618,6 +3618,12 @@ abstract class S {
   /// In en, this message translates to:
   /// **'Sorting unavailable during text search'**
   String get browseSortDisabledHint;
+
+  /// No description provided for @typeToFilterHint.
+  ///
+  /// In en, this message translates to:
+  /// **'Filter...'**
+  String get typeToFilterHint;
 }
 
 class _SDelegate extends LocalizationsDelegate<S> {